requirements.txt
fastapi==0.110.0
uvicorn==0.27.0
python-multipart==0.0.6
pydantic==2.5.2
pdfplumber==0.10.2
python-docx==1.0.1
pytesseract==0.3.10
pillow==10.1.0
python-dotenv==1.0.0
openai==1.13.3
langchain==0.0.335
langchain-openai==0.0.5
langgraph==0.0.26
